Migrate modals.js to modules. Moved checkLeave to editor.js. Removed modals.js reference in the html page.
