+++ blocklyc.html
@@ -98,7 +98,6 @@
     <script type="text/javascript" src="src/propterm.js" defer></script>
     <script type="text/javascript" src="src/blocklypropclient.js" defer></script>
     <script type="text/javascript" src="src/blocklyc.js" defer></script>
-    <script type="text/javascript" src="src/modals.js" defer></script>
     <script type="text/javascript" src="src/editor_js.js" defer></script>
     <!-- <script type="text/javascript" src="src/localcompile.js"></script> -->
 
@@ -127,8 +126,9 @@
     <!-- Module Conversion -->
     <script type="text/javascript" src="src/editor_js.js" defer></script>
 
-<!--    <script type="module" src="src/modules/project_timer.js" defer></script> -->
-<!--    <script type="module" src="src/modules/toolbar_controller.js"></script>  -->
+<!--    <script type="text/javascript" src="src/modules/modals.js" defer></script> -->
+<!--    <script type="module" src="src/modules/project_timer.js" defer></script>   -->
+<!--    <script type="module" src="src/modules/toolbar_controller.js"></script>    -->
     <script type="module" src="src/modules/editor.js" defer></script>
 
     <!-- Style sheets -->
